+++ src/geo/mod.rs
@@ -1,4 +1,5 @@
 pub mod datatypes;
 pub mod functions;
+pub mod gml;
 pub mod spatial_index;
 pub mod vocabulary;